Use standard as our Ruby linter
Installing Niftany, our current linter, posed some problems because of
its dependency on erb-linter. Since this probject isn't Rails, using
Niftany will not work. As an alternative, we can use the standard gem.

The rules differ from our Niftany gem, but the advantage is there is no
need for shared Rubocop rules. Simply install the gem and run the task.