Add additional formats when microfilm already exists (#306)
* Processor::PreferredFormat

There is a "preferred" format that we can determine first, and if it's
found, return early to avoid additional processing.

* Use Processors::RecordType

Extracts the logic for determining record type for a format into a
separate class. This class is responsible for adding additional formats
if the exiting formats include a local microfilm format type.

* Refactor PreferredFormat

This processor can now return a string, and the logic for determining
that is simplified.

* Add testing coverage for instructional materials

These didn't appear be covered in the original spec tests, so we're
adding specific tests that look at the 006 and 008 fields for that
format.